project/02-pjt/movies/permmisions.py
#permission.py
from rest_framework import permissions

class IsAuthorOrReadOnly(permissions.BasePermission):
    """
    SAFE_METHODS(GET, HEAD, OPTIONS)은 허용
    PUT, PATCH, DELETE는 작성자(author)만 허용
    """
    def has_object_permission(self, request, view, obj):
        if request.method in permissions.SAFE_METHODS:
            return True
        return obj.author == request.user
